backend-bcr/requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.1
black==22.12.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==39.0.0
cycler==0.11.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
Flask==2.1.1
fonttools==4.32.0
google==3.0.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.13.2
google-cloud-core==2.3.1
google-cloud-speech==2.14.1
google-cloud-vision==2.7.3
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.4
httptools==0.5.0
httpx==0.23.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
numpy==1.22.3
orjson==3.8.5
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
proto-plus==1.20.6
protobuf==3.20.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.28.0
rfc3986==1.5.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.1.1
XlsxWriter==3.0.8
